WARMUP/ITER + GFLOPS + modif mmio.c/h

+++ GPU_solution/src/my_time_lib.c
@@ -1,34 +1,34 @@
 #include "../include/my_time_lib.h"
 
 
-double arithmetic_mean(double *v, int len) {
+float arithmetic_mean(float *v, int len) {
 
-    double mu = 0.0;
+    float mu = 0.0;
     for (int i=0; i<len; i++)
-        mu += (double)v[i];
-    mu /= (double)len;
+        mu += (float)v[i];
+    mu /= (float)len;
 
     return(mu);
 }
 
-double geometric_mean(double *v, int len) {
+float geometric_mean(float *v, int len) {
     
-    double mu = 1.0;
+    float mu = 1.0;
     for (int i=0; i<len; i++) {
-        mu *= (v[i] > 0) ? ((double)v[i]) : 1;
+        mu *= (v[i] > 0) ? ((float)v[i]) : 1;
     }
     mu = pow(mu, 1.0 / len);
     
     return(mu);
 }
 
-double sigma_fn_sol(double *v, double mu, int len) {
+float sigma_fn_sol(float *v, float mu, int len) {
 
-    double sigma = 0.0;
+    float sigma = 0.0;
     for (int i=0; i<len; i++) {
-        sigma += ((double)v[i] - mu)*((double)v[i] - mu);
+        sigma += ((float)v[i] - mu)*((float)v[i] - mu);
     }
-    sigma /= (double)len;
+    sigma /= (float)len;
 
     return(sigma);
 }
